tests: fix tests
Moved and extended the should_skip function + fixed some tests
Enable parallel execution which removes the need to cleanup things
after a test.

Moved the ACL tests in a specific file.
Fixtures are now function scoped. This *truly* enables
multi-version tests.
