+++ tox.ini
@@ -5,7 +5,9 @@ envlist = lint,py38,py39,py10,py11,py12
 deps =
     -r tests-requirements.txt
 commands =
-    py.test --reruns=3 {posargs:consul tests}
+    coverage erase
+    pytest -nauto --reruns=3 tests
+    coverage report
 
 # Linter environment
 [testenv:lint]
